app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic-settings==2.7.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pyodbc==5.3.0
azure-monitor-opentelemetry==1.8.7
azure-appconfiguration-provider==2.4.0
azure-identity==1.25.3
azure-storage-blob==12.28.0
